extract db  out of test context
