+++ aptos-indexer-processors-sdk/testing-framework/src/lib.rs
@@ -1,2 +1,3 @@
 mod mock_grpc;
 pub mod new_test_context;
+pub mod database;
