+++ aptos-indexer-processors-sdk/testing-framework/src/new_test_context.rs
@@ -1,36 +1,32 @@
+use std::error::Error;
 use std::fs;
-use std::path::Path;
+use std::path::{Path, PathBuf};
+use anyhow::Context;
 use aptos_protos::indexer::v1::TransactionsResponse;
 use aptos_indexer_processor_sdk::traits::processor_trait::ProcessorTrait;
 use aptos_protos::transaction::v1::Transaction;
-use async_trait::async_trait;
-use testcontainers::{
-    core::{IntoContainerPort, WaitFor},
-    runners::AsyncRunner,
-    ContainerAsync, GenericImage, ImageExt,
-};
+use serde_json::to_string_pretty;
+
 use url::Url;
 use aptos_indexer_processor_sdk::aptos_indexer_transaction_stream::TransactionStreamConfig;
 use crate::mock_grpc::MockGrpcServer;
 
-pub struct SdkTestContext<D: TestDatabase> {
+const DEFAULT_OUTPUT_FOLDER: &str = "expected_db_output_files/";
+const INDEXER_GRPC_DATA_SERVICE_URL: &str = "http://localhost:51254";
+
+pub struct SdkTestContext{
     pub transaction_batches: Vec<Transaction>,
-    pub database: D, // Holds the database setup (user-defined)
 }
 
-impl<D: TestDatabase> SdkTestContext<D> {
-    pub async fn new(txn_bytes: &[&[u8]], mut database: D) -> anyhow::Result<Self> {
+impl SdkTestContext {
+    pub async fn new(txn_bytes: &[&[u8]]) -> anyhow::Result<Self> {
         let transaction_batches = txn_bytes
             .iter()
             .map(|txn| serde_json::from_slice(txn).unwrap())
             .collect::<Vec<Transaction>>();
 
-        // Set up the database using the user-defined method
-        database.setup().await?;
-
         Ok(SdkTestContext {
             transaction_batches,
-            database,
         })
     }
 
@@ -38,7 +34,8 @@ impl<D: TestDatabase> SdkTestContext<D> {
     pub async fn run<F>(
         &self,
         processor: &impl ProcessorTrait,
-        txn_verion: u64,
+        db_url: &str,
+        txn_version: u64,
         generate_files: bool, // flag to control file generation
         output_path: Option<String>, // Optional custom output path
         verification_f: F,
@@ -57,32 +54,16 @@ impl<D: TestDatabase> SdkTestContext<D> {
         // Run the processor
         processor.run_processor().await?;
 
-        // Pass the DB URL for user-defined validation
-        let db_url = self.database.get_db_url();
-        let db_values = verification_f(&db_url)?;
+        let db_values = verification_f(db_url)?;
 
         // Conditionally generate output files if the `generate_files` flag is true
         if generate_files {
-            // Use the provided output path or fall back to default
-            let processor_name = processor.name();
-            let processor_name = "events_processor"; //  TODO: add a function to get name in runtime
-            let txn_version = txn_verion.to_string();
-            let output_dir = output_path.unwrap_or_else(|| {
-                format!(
-                    "expected_db_output_files/imported_testnet_txns/{}/{}_{}.json",
-                    processor_name, processor_name, txn_version
-                )
-            });
-
-            // Extract the directory path from the full file path
-            if let Some(parent_dir) = Path::new(&output_dir).parent() {
-                // Create the directory if it doesn't exist
-                fs::create_dir_all(parent_dir)?;
-            }
-
-            // Save the database output as JSON to the specified path
-            fs::write(output_dir.clone(), serde_json::to_string_pretty(&db_values)?)?;
-            println!("[INFO] Generated output file at: {}", output_dir);
+            generate_output_file(
+                processor.name(),
+                &txn_version.to_string(),
+                &db_values,
+                output_path
+            )?;
         } else {
             println!("[INFO] Skipping file generation as requested.");
         }
@@ -110,12 +91,12 @@ impl<D: TestDatabase> SdkTestContext<D> {
         ending_version: u64,
     ) -> TransactionStreamConfig {
         TransactionStreamConfig {
-            indexer_grpc_data_service_address: Url::parse("http://localhost:51254")
+            indexer_grpc_data_service_address: Url::parse(INDEXER_GRPC_DATA_SERVICE_URL)
                 .expect("Could not parse database url"),
             starting_version: Some(starting_version), // dynamically pass the starting version
             request_ending_version: Some(ending_version), // dynamically pass the ending version
             auth_token: "".to_string(),
-            request_name_header: "sdk testing".to_string(),
+            request_name_header: "sdk-testing".to_string(),
             indexer_grpc_http2_ping_interval_secs: 30,
             indexer_grpc_http2_ping_timeout_secs: 10,
             indexer_grpc_reconnection_timeout_secs: 10,
@@ -125,72 +106,35 @@ impl<D: TestDatabase> SdkTestContext<D> {
 
 }
 
-#[async_trait]
-pub trait TestDatabase: Send + Sync {
-    /// Set up the test container using user-defined code.
-    async fn setup<'a>(&'a mut self) -> anyhow::Result<()>;
-
-    /// Retrieve the database connection URL after setup.
-    fn get_db_url(&self) -> String;
+// Helper function to construct the output file path
+fn construct_file_path(output_dir: &str, processor_name: &str, txn_version: &str) -> PathBuf {
+    Path::new(output_dir)
+        .join(processor_name)
+        .join(format!("{}_{}.json", processor_name, txn_version))
 }
 
-pub struct PostgresTestDatabase {
-    connection_string: String,
-    postgres_container: Option<ContainerAsync<GenericImage>>,
-}
-
-impl PostgresTestDatabase {
-    pub fn new() -> Self {
-        PostgresTestDatabase {
-            postgres_container: None,
-            connection_string: String::new(),
-        }
+// Helper function to ensure the directory exists
+fn ensure_directory_exists(path: &Path) -> anyhow::Result<()> {
+    if let Some(parent_dir) = path.parent() {
+        fs::create_dir_all(parent_dir).expect("Failed to create directory");
     }
+    Ok(())
 }
 
-#[async_trait]
-impl TestDatabase for PostgresTestDatabase {
-    /// Set up the Postgres container and get the database connection URL.
-    async fn setup<'a>(&'a mut self) -> anyhow::Result<()> {
-        self.postgres_container = Some(
-            GenericImage::new("postgres", "14")
-                .with_exposed_port(5432.tcp())
-                .with_wait_for(WaitFor::message_on_stderr(
-                    "database system is ready to accept connections",
-                ))
-                .with_env_var("POSTGRES_DB", "postgres")
-                .with_env_var("POSTGRES_USER", "postgres")
-                .with_env_var("POSTGRES_PASSWORD", "postgres")
-                .start()
-                .await
-                .expect("Postgres started"),
-        );
+// Helper function to generate output files
+fn generate_output_file(
+    processor_name: &str,
+    txn_version: &str,
+    db_values: &serde_json::Value,
+    output_path: Option<String>,
+) -> anyhow::Result<()> {
+        let output_dir = output_path.unwrap_or_else(|| DEFAULT_OUTPUT_FOLDER.to_string());
+        let file_path = construct_file_path(&output_dir, processor_name, txn_version);
 
-        // Retrieve the host and port of the container for the connection string
-        let host = self
-            .postgres_container
-            .as_ref()
-            .unwrap()
-            .get_host()
-            .await
-            .expect("Failed to get host");
+        ensure_directory_exists(&file_path)?;
 
-        let port = self
-            .postgres_container
-            .as_ref()
-            .unwrap()
-            .get_host_port_ipv4(5432)
-            .await
-            .expect("Failed to get port");
-
-        // Create the Postgres connection string
-        self.connection_string = format!("postgres://postgres:postgres@{}:{}/postgres", host, port);
-
+        fs::write(&file_path, to_string_pretty(db_values)?)
+            .context(format!("Failed to write file to {:?}", file_path))?;
+        println!("[INFO] Generated output file at: {}", file_path.display());
         Ok(())
-    }
-
-    /// Retrieve the Postgres connection URL after the container has been set up.
-    fn get_db_url(&self) -> String {
-        self.connection_string.clone()
-    }
 }
