+++ aptos-indexer-processors-sdk/transaction-stream/Cargo.toml
@@ -18,7 +18,6 @@ aptos-moving-average = { workspace = true }
 aptos-protos = { workspace = true }
 chrono = { workspace = true }
 futures-util = { workspace = true }
-mockall = { workspace = true }
 once_cell = { workspace = true }
 prometheus = { workspace = true }
 prost = { workspace = true }
@@ -28,4 +27,3 @@ tokio = { workspace = true }
 tonic = { workspace = true }
 tracing = { workspace = true }
 url = { workspace = true }
-async-trait = { workspace = true }
